% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocessing.R
\name{cleaning_df}
\alias{cleaning_df}
\title{This function "cleans" up a dataframe so that it can be used by the application}
\usage{
cleaning_df(data)
}
\arguments{
\item{data}{A dataframe}
}
\value{
a dataframe that can be used by the application
}
\description{
This function "cleans" up a dataframe so that it can be used by the application
}
\details{
For the application to work well, the input dataframe data should
contain only numeric variables with no missing values.
This function acts as a safeguard in case the user does not clean up his dataframe.
Qualitative variables are removed and only non-missing values are kept
in the final dataframe.
}
\author{
Josué AFOUDA
}
