% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocessing.R
\name{normalize}
\alias{normalize}
\title{Function to normalize a variable by shift the scale so that it starts at O,
and then compress it so that it ends on 1.}
\usage{
normalize(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector of numeric values}

\item{na.rm}{A boolean value TRUE or FALSE. The default is set to TRUE}
}
\value{
a vector contains normalized values
}
\description{
Function to normalize a variable by shift the scale so that it starts at O,
and then compress it so that it ends on 1.
}
\details{
THis function takes two arguments, a vector x of numeric values,
and the logical na.rm and return a vector of normalized values
}
\examples{
# Normalize a single vector
normalize(iris$Sepal.Length)
# Normalize all variables of a dataframe
normalize(mtcars)
}
\author{
Josué AFOUDA
}
