% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{sampleData}
\alias{sampleData}
\title{Iris dataset without the Species column}
\format{
A data frame with 150 rows and 4 variables
}
\source{
Created as a sample
}
\usage{
sampleData
}
\description{
Iris dataset without the Species column
}
\details{
iris is a data frame with 150 cases (rows) and 5 variables (columns) named
Sepal.Length, Sepal.Width, Petal.Length, Petal.Width, and Species.

\itemize{
  \item Sepal.Length. Sepal length measurements.
  \item Sepal.Width. Sepal width measurements.
  \item Petal.Length. Petal length measurements.
  \item Petal.Width. Pepal width measurements.
}
}
\keyword{datasets}
