% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyul.R
\name{ShinyUL}
\alias{ShinyUL}
\title{Function to run the web application}
\usage{
ShinyUL(input_data, seed = 123)
}
\arguments{
\item{input_data}{A dataframe}
}
\value{
a shiny web application
}
\description{
Function to run the web application
}
\details{
This function takes one argument, a dataframe, and
returns a shiny web app that allows various unsupervised learning
techniques to be applied to a dataframe
}
\author{
Josué AFOUDA
}
