% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocessing.R
\name{standardize}
\alias{standardize}
\title{Function to standardize a variable. After being standardized,
the variable has zero mean and unit standard deviation.}
\usage{
standardize(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector of numeric values}
}
\value{
a vector contains standard scores a.k.a z-scores
}
\description{
Function to standardize a variable. After being standardized,
the variable has zero mean and unit standard deviation.
}
\details{
THis function takes one argument, a vector x of numeric values.
and return a vector of z-scores that represent the number of standard deviations
above or below the mean that a specific observation falls.
}
\examples{
# Example usage of standardize
standardize(iris$Sepal.Length)
standardize(mtcars)
}
\author{
Josué AFOUDA
}
